Fix change sim

See merge request !151

+++ app/src/main/java/com/viettel/mbccs/data/source/remote/request/GetListDsLamByTeamIdRequest.java
@@ -5,5 +5,5 @@ package com.viettel.mbccs.data.source.remote.request;
  */
 
 public class GetListDsLamByTeamIdRequest {
-    public Long shopId;
+    public int shopId;
 }
